DEV |Recommandation
 - Ajout Soumission de recommandation dans les biens
- correction petits bug liste Archived

+++ src/Service/ArchivePropertyService.php
@@ -5,6 +5,7 @@ namespace App\Service;
 
 use App\Entity\Gestapp\Property;
 use App\Entity\Gestapp\Transaction;
+use App\Repository\Admin\ContactRepository;
 use App\Repository\Gestapp\CadasterRepository;
 use App\Repository\Gestapp\ComplementRepository;
 use App\Repository\Gestapp\PhotoRepository;
@@ -16,7 +17,13 @@ use Symfony\Component\HttpFoundation\RequestStack;
 class ArchivePropertyService{
 
     public function __construct(
-        public PropertyRepository $propertyRepository
+        public PropertyRepository $propertyRepository,
+        public PhotoRepository $photoRepository,
+        public CadasterRepository $cadasterRepository,
+        public PublicationRepository $publicationRepository,
+        public ComplementRepository $complementRepository,
+        public TransactionRepository $transactionRepository,
+        public ContactRepository $contactRepository
     )
     {}
 
@@ -42,26 +49,41 @@ class ArchivePropertyService{
         CadasterRepository $cadasterRepository,
         PublicationRepository $publicationRepository,
         ComplementRepository $complementRepository,
-        TransactionRepository $transactionRepository
+        TransactionRepository $transactionRepository,
+        ContactRepository $contactRepository
     )
     {
         // Suppression des entités liées à la propriété
         $publication = $property->getPublication();
         $complement = $property->getOptions();
         // 1.Supression des images liées à la propriété
-        $photos = $photoRepository->findBy(['property' => $property]);
-        foreach($photos as $photo){
-            $photoRepository->remove($photo);
+        $photos = $this->photoRepository->findBy(['property' => $property]);
+        //dd($photos);
+        if(count($photos) > 0){
+            foreach($photos as $photo){
+                $photoRepository->remove($photo);
+            }
         }
         // 2.supression des zones de cadastres liées à la propriété
-        $cadasters = $cadasterRepository->findBy(['property' => $property]);
-        foreach($cadasters as $cadaster){
-            $cadasterRepository->remove($cadaster);
+        $cadasters = $this->cadasterRepository->findBy(['property' => $property]);
+        //dd('cadasters' , $cadasters);
+        if(count($cadasters) > 0){
+            foreach($cadasters as $cadaster){
+                $cadasterRepository->remove($cadaster);
+            }
         }
-        $transactions = $transactionRepository->findBy(['property' => $property]);
-        if($transactions){
+        $transactions = $this->transactionRepository->findBy(['property' => $property]);
+        //dd('transaction' , $transactions);
+        if(count($transactions) > 0){
             foreach($transactions as $t){
                 $property->removeTransaction($t);
+            }
+        }
+        $contacts = $this->contactRepository->findBy(['property' => $property]);
+        //dd('contacts' , $contacts);
+        if(count($contacts) > 0){
+            foreach($contacts as $c){
+                $this->contactRepository->remove($c);
             }
         }
         // 3. Finalisation des suppression
